Move JShell to its own package
Let's see how nicely git tracks the rename...
